combined_tests/requirements.txt
pyqt5==5.12.3
pyserial
pyinstaller
pyusb
